+++ src/SoundFingerprinting/IModelService.cs
@@ -9,9 +9,9 @@ namespace SoundFingerprinting
 
     public interface IModelService
     {
-        IList<SubFingerprintData> ReadSubFingerprints(long[] hashBins, QueryConfiguration config);
+        IList<SubFingerprintData> ReadSubFingerprints(int[] hashBins, QueryConfiguration config);
 
-        ISet<SubFingerprintData> ReadSubFingerprints(IEnumerable<long[]> hashes, QueryConfiguration config);
+        ISet<SubFingerprintData> ReadSubFingerprints(IEnumerable<int[]> hashes, QueryConfiguration config);
 
         IModelReference InsertTrack(TrackData track);
