+++ src/SoundFingerprinting/Math/ISimilarityUtility.cs
@@ -14,7 +14,7 @@ namespace SoundFingerprinting.Math
 
         int CalculateHammingSimilarity(byte[] a, byte[] b);
 
-        int CalculateHammingSimilarity(long[] expected, long[] actual, int setBytesPerLong);
+        int CalculateHammingSimilarity(int[] expected, int[] actual, int setBytesPerLong);
 
         /// <summary>
         ///   Calculate similarity between 2 fingerprints.
